pyproject.toml
[project]
name = "coincall-rfq-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.1",
    "websockets>=15.0.1",
]
